added EventTracker in Manager and Custom signal System to Connect

+++ Assets/HelloPico2/Event Tracker/EventExporter.cs
@@ -1,28 +1,38 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using UnityEngine;
 
 namespace Actor.Scripts.EventMessage{
 	public class EventExporter{
 		public string FileName{ get; }
-
+		public string FileFullPath{ get; }
 		private readonly FileStream _file;
-
-		private string _fileFullName;
-
-		private StreamWriter _streamWriter;
+		private readonly StreamWriter _streamWriter;
 
 		public EventExporter(string fileName){
-			_fileFullName = Application.persistentDataPath + "/" + $"{fileName}" + ".json";
-			_file = new FileStream(_fileFullName, FileMode.CreateNew);
+			FileFullPath = Application.persistentDataPath + "/" + $"{fileName}" + ".json";
+			if(File.Exists(FileFullPath)) //TODO : handle this situation
+				return;
+
+			_file = new FileStream(FileFullPath, FileMode.CreateNew);
 			_streamWriter = new StreamWriter(_file);
 			FileName = fileName;
 		}
 
 
-		public void WriteEvent(){
-			// var jsonString = JsonUtility.ToJson();
-			// _streamWriter.Write(jsonString + Environment.NewLine);
+		public void WriteEvent(object eventData){
+			var jsonString = GetJsonString(eventData);
+			_streamWriter.Write(jsonString + Environment.NewLine);
 			_streamWriter.Flush();
+		}
+
+		public string GetJsonString(object eventData){
+			var type = eventData.GetType();
+			if(!type.IsSerializable){
+				throw new Exception($"{type} is not Serializable Type");
+			}
+			var jsonString = JsonUtility.ToJson(eventData);
+			return jsonString;
 		}
 
 		public void Timeout(){
@@ -30,10 +40,10 @@ namespace Actor.Scripts.EventMessage{
 			_streamWriter.Close();
 			_streamWriter.Dispose();
 		}
-		public static void PostTrackerEvent(string eventName,string jsonString) 
-		{
+
+		public static void PostTrackerEvent(string eventName, string jsonString){
 			var cls = new AndroidJavaClass("os.teatracker.TeaAgent");
-			var jsonParam = new AndroidJavaObject("org.json.JSONObject",jsonString);
+			var jsonParam = new AndroidJavaObject("org.json.JSONObject", jsonString);
 			cls.CallStatic("onEvent", eventName, jsonParam);
 		}
